utils/event_research.py
import os
import logging
from serpapi import GoogleSearch  # placeholder for event API

logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)

# Placeholder Event Discovery: searches for events by keyword via SERP API
SERP_KEY = os.getenv("SERPAPI_API_KEY")

def discover_events(keyword: str, max_events: int = 5) -> list:
    """
    Dynamic Event Research: Given a keyword, return a list of event URLs.
    This is a placeholder for a true event API (Eventbrite, Meetup, etc.).
    """
    if not SERP_KEY:
        logger.error("SERPAPI_API_KEY not set for event discovery")
        return []
    params = {"engine":"google","q":f"{keyword} trade show 2025 site:mapyourshow.com","api_key":SERP_KEY}
    try:
        results = GoogleSearch(params).get_dict().get("organic_results", [])
    except Exception as e:
        logger.exception("Event discovery failed")
        return []
    urls = []
    for r in results:
        link = r.get("link")
        if link and link not in urls:
            urls.append(link)
        if len(urls) >= max_events:
            break
    logger.info(f"Discovered {len(urls)} event URLs for keyword '{keyword}'")
    return urls
